@@ -7,49 +7,51 @@
   "devDependencies": {
     "@babel/core": "7.12.13",
     "@babel/preset-typescript": "7.12.13",
-    "@nrwl/cli": "13.8.1",
-    "@nrwl/eslint-plugin-nx": "13.8.1",
-    "@nrwl/jest": "13.8.1",
-    "@nrwl/js": "13.8.1",
-    "@nrwl/linter": "13.8.1",
-    "@nrwl/react": "^13.8.1",
-    "@nrwl/storybook": "^13.8.1",
-    "@nrwl/tao": "13.8.1",
-    "@nrwl/web": "13.8.1",
-    "@nrwl/workspace": "13.8.1",
-    "@storybook/addon-essentials": "~6.4.12",
-    "@storybook/builder-webpack5": "~6.4.12",
-    "@storybook/manager-webpack5": "~6.4.12",
-    "@storybook/react": "~6.4.12",
+    "@nrwl/cli": "14.5.10",
+    "@nrwl/eslint-plugin-nx": "14.5.10",
+    "@nrwl/jest": "14.5.10",
+    "@nrwl/js": "14.5.10",
+    "@nrwl/linter": "14.5.10",
+    "@nrwl/react": "14.5.10",
+    "@nrwl/storybook": "14.5.10",
+    "@nrwl/web": "14.5.10",
+    "@nrwl/workspace": "14.5.10",
+    "@storybook/react": "6.5.10",
+    "@storybook/addon-essentials": "6.5.10",
+    "@storybook/builder-webpack5": "6.5.10",
+    "@storybook/manager-webpack5": "6.5.10",
     "@svgr/webpack": "^5.4.0",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/react-hooks": "7.0.2",
-    "@types/jest": "27.0.2",
-    "@types/node": "16.11.7",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
-    "babel-jest": "27.2.3",
+    "@testing-library/react": "13.3.0",
+    "@types/jest": "27.4.1",
+    "@types/node": "18.7.5",
+    "@types/react": "18.0.17",
+    "@types/react-dom": "18.0.6",
+    "@typescript-eslint/eslint-plugin": "5.35.1",
+    "@typescript-eslint/parser": "5.35.1",
+    "babel-jest": "27.5.1",
     "babel-loader": "8.1.0",
-    "eslint": "~8.7.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-import": "2.25.2",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.27.0",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "jest": "27.2.3",
-    "prettier": "^2.5.1",
-    "react-test-renderer": "17.0.2",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.5.2",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "jest": "27.5.1",
+    "nx": "14.5.10",
+    "prettier": "2.7.1",
+    "react-test-renderer": "18.2.0",
+    "ts-jest": "27.1.4",
+    "ts-node": "10.9.1",
+    "typescript": "4.7.4",
     "url-loader": "^3.0.0"
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-is": "18.0.0",
     "regenerator-runtime": "0.13.7",
-    "tslib": "^2.0.0"
+    "tslib": "^2.0.0",
+    "@storybook/core-server": "^6.5.10"
   }
 }
