@@ -1,10 +1,10 @@
-import React from "react";
+import React, {PropsWithChildren} from "react";
 import {IButtonGroupProps} from "./button-group.props.interface";
 
 import styles from './button-group.module.scss';
 import classNames from "classnames";
 
-export const ButtonGroup: React.FC<IButtonGroupProps> = ({children, className}) => {
+export const ButtonGroup: React.FC<PropsWithChildren<IButtonGroupProps>> = ({children, className}) => {
   return (
     <div className={classNames(styles.btn_group, className)}>
       {children}
