@@ -1,9 +1,9 @@
-import React from "react";
+import React, {PropsWithChildren} from "react";
 import {IRadioGroup} from "./radio-group.props.interface";
 
 import styles from './radio-group.module.scss';
 
-export const RadioGroup: React.FC<IRadioGroup> = props => {
+export const RadioGroup: React.FC<PropsWithChildren<IRadioGroup>> = props => {
   const [selectedIndex, setSelectedIndex] = React.useState(0);
   return (
     <label className={styles.label}>{props.label}
