@@ -1,10 +1,9 @@
 import { render } from '@testing-library/react';
-
-import TextField from './textfield';
+import TextField from "./text-field";
 
 describe('Textfield', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<TextField />);
+    const { baseElement } = render(<TextField style="default" />);
     expect(baseElement).toBeTruthy();
   });
 });
